refactor: split warehouse page into tab components
Extract Pick/Prep, Deploy, and Return tabs into dedicated component files.
Main page reduced from 3366 to 2395 lines. Shared types and utilities
moved to warehouse-types.ts. KitChildRows and PrepStatusBadge extracted
as shared components used across multiple tabs. Pure prop-passing, no
new abstractions or context providers.
